# docs/openapi/repository-api-v1.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "description": "Error detail object for the public API error envelope.",
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured metadata to help clients debug the error.",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable summary of the failure.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "Standardized error envelope shared across HTTP endpoints.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody",
            "description": "Normalized error payload."
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "SkillSearchResponse": {
        "description": "Compact advisory search response.",
        "properties": {
          "results": {
            "description": "One compact candidate per skill identifier in deterministic order.",
            "items": {
              "$ref": "#/components/schemas/SkillSearchResultResponse"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "SkillSearchResponse",
        "type": "object"
      },
      "SkillSearchResultResponse": {
        "description": "Compact advisory candidate returned by the search API.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional short skill description.",
            "title": "Description"
          },
          "footprint_bytes": {
            "description": "Artifact size used for compact filtering and ranking.",
            "title": "Footprint Bytes",
            "type": "integer"
          },
          "freshness_days": {
            "description": "Whole days since the version was published.",
            "title": "Freshness Days",
            "type": "integer"
          },
          "matched_fields": {
            "description": "Fields that contributed to the candidate match explanation.",
            "items": {
              "type": "string"
            },
            "title": "Matched Fields",
            "type": "array"
          },
          "matched_tags": {
            "description": "Normalized tag filters satisfied by this result.",
            "items": {
              "type": "string"
            },
            "title": "Matched Tags",
            "type": "array"
          },
          "name": {
            "description": "Human-readable skill name.",
            "title": "Name",
            "type": "string"
          },
          "published_at": {
            "description": "UTC timestamp when the version was published.",
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "reasons": {
            "description": "Stable advisory ranking hints describing why the result matched.",
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "skill_id": {
            "description": "Stable identifier of the matched skill.",
            "title": "Skill Id",
            "type": "string"
          },
          "tags": {
            "description": "Searchable tags attached to the matched version.",
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "usage_count": {
            "description": "Derived usage metric reserved for future ranking inputs.",
            "title": "Usage Count",
            "type": "integer"
          },
          "version": {
            "description": "Best matching immutable version for this skill.",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "skill_id",
          "version",
          "name",
          "description",
          "tags",
          "published_at",
          "freshness_days",
          "footprint_bytes",
          "usage_count",
          "matched_fields",
          "matched_tags",
          "reasons"
        ],
        "title": "SkillSearchResultResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Registry-first API for immutable skill publication, indexed search candidate\nretrieval, direct relationship reads, exact version retrieval, artifact streaming,\nand version listing.\n\nThe server owns data-local registry operations only. Prompt interpretation,\nreranking, dependency solving, final selection, lock generation, and execution\nplanning remain client-owned behavior outside this API boundary.",
    "title": "aptitude-server Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/discovery/skills/search": {
      "get": {
        "description": "Return compact advisory candidates from indexed metadata and description search. This route is discovery-only: it does not resolve relationships, choose a final candidate, or perform dependency solving.",
        "operationId": "searchDiscoverySkills",
        "parameters": [
          {
            "description": "Full-text discovery query.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Full-text discovery query.",
              "title": "Q"
            }
          },
          {
            "description": "Repeatable tag filter; all values must match.",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeatable tag filter; all values must match.",
              "title": "Tag"
            }
          },
          {
            "description": "Language alias filter implemented as a normalized tag.",
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Language alias filter implemented as a normalized tag.",
              "title": "Language"
            }
          },
          {
            "description": "Maximum age in days since publication.",
            "in": "query",
            "name": "fresh_within_days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum age in days since publication.",
              "title": "Fresh Within Days"
            }
          },
          {
            "description": "Maximum allowed artifact size in bytes.",
            "in": "query",
            "name": "max_footprint_bytes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum allowed artifact size in bytes.",
              "title": "Max Footprint Bytes"
            }
          },
          {
            "description": "Maximum number of skill candidates to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Maximum number of skill candidates to return.",
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "results": [
                    {
                      "description": "Linting skill",
                      "footprint_bytes": 123,
                      "freshness_days": 0,
                      "matched_fields": [
                        "name",
                        "tags"
                      ],
                      "matched_tags": [
                        "python"
                      ],
                      "name": "Python Lint",
                      "published_at": "2026-03-10T08:30:00Z",
                      "reasons": [
                        "text_match",
                        "tag_filter_match"
                      ],
                      "skill_id": "python.lint",
                      "tags": [
                        "python",
                        "lint"
                      ],
                      "usage_count": 0,
                      "version": "1.2.3"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/SkillSearchResponse"
                }
              }
            },
            "description": "Compact advisory skill candidates returned successfully."
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "invalid_request": {
                    "value": {
                      "error": {
                        "code": "INVALID_REQUEST",
                        "details": {
                          "errors": [
                            {
                              "input": "latest",
                              "loc": [
                                "path",
                                "version"
                              ],
                              "msg": "String should match pattern",
                              "type": "string_pattern_mismatch"
                            }
                          ]
                        },
                        "message": "Request validation failed."
                      }
                    }
                  },
                  "missing_selector": {
                    "value": {
                      "error": {
                        "code": "INVALID_REQUEST",
                        "details": {
                          "errors": [
                            {
                              "input": {
                                "limit": 20,
                                "tags": []
                              },
                              "loc": [],
                              "msg": "Value error, At least one search selector must be provided.",
                              "type": "value_error"
                            }
                          ]
                        },
                        "message": "Search request validation failed."
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The discovery request is invalid."
          }
        },
        "summary": "Search skill discovery metadata",
        "tags": [
          "discovery"
        ]
      }
    },
...
  }
}